requirements.txt
# Core Deep Learning
torch>=1.12.0
torchvision>=0.13.0
timm>=0.9.0

# Computer Vision
opencv-python>=4.7.0
Pillow>=9.5.0

# Data Processing
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
PyYAML>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=7.2.0
pytest-cov>=4.0.0

# Logging and Experiment Tracking
tensorboard>=2.12.0

# Optional (uncomment if using Weights & Biases)
# wandb>=0.15.0

# Optional (uncomment if using Detectron2 for Mask R-CNN)
# detectron2 - install separately: python -m pip install 'git+https://github.com/facebookresearch/detectron2.git'

# Kaggle API (for dataset download)
kaggle>=1.5.16

# Data augmentation
albumentations>=1.3.0

# Additional utilities
pycocotools>=2.0.6
scipy>=1.10.0
